+++ tests/misc.js
@@ -74,40 +74,16 @@ test('snapshot option', function (t) {
 })
 
 test('swarm options', function (t) {
-  var dat = Dat({utp: false, port: 1234, discovery: {upload: false}, webrtc: false, dir: process.cwd()})
+  var dat = Dat({utp: false, port: 1234, discovery: {upload: false}, dir: process.cwd()})
   dat.open(function (err) {
     t.error(err)
     dat.once('connecting', function () {
       var swarm = dat.swarm
 
-      t.ok(!swarm.opts.discovery.utp, 'utp discovery false')
-      t.ok(!swarm.node._utp, 'No utp discovery')
+      t.ok(!swarm._utp, 'No utp discovery')
 
-      t.same(swarm.opts.port, 1234, 'port option set on swarm')
-      t.same(swarm.node.address().port, 1234, 'port on node swarm okay')
+      t.same(swarm.address().port, 1234, 'port on swarm okay')
 
-      t.ok(!swarm.uploading, 'Upload false set on swarm')
-
-      t.ok(!swarm.opts.wrtc, 'Swarm webrtc option false')
-
-      dat.close(function () {
-        dat.db.close(function () {
-          rimraf(path.join(process.cwd(), '.dat'), function () {
-            t.end()
-          })
-        })
-      })
-    })
-    dat._joinSwarm()
-  })
-})
-
-test('swarm options 3.2.x compat', function (t) {
-  var dat = Dat({upload: false, dir: process.cwd()})
-  dat.open(function (err) {
-    t.error(err)
-    dat.once('connecting', function () {
-      var swarm = dat.swarm
       t.ok(!swarm.uploading, 'Upload false set on swarm')
 
       dat.close(function () {