+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- Change types:
   ### Added, ### Changed, ### Fixed, ### Removed, ### Deprecated
 -->
+### Removed
+* `webrtc` option.
+* `opts.upload` no longer supported option, use `opts.discovery.upload`.
 
 ## 3.6.0 - 2016-09-27
 ### Added